+++ PasteBinIngest.Data/Constants.cs
@@ -1,14 +1,17 @@
 ﻿namespace PasteBinIngest.Data
 {
+    /// <summary>
+    /// constants used in the repositories.
+    /// </summary>
     internal class Constants
     {
         // Directory Names
-        public const string RequestDirectory = "/requests";
-        public const string EntryDirectory = "/entries";
+        public const string RequestDirectory = "\\requests";
+        public const string EntryDirectory = "\\entries";
 
         // Entry File Names
-        public const string Raw = "raw";
-        public const string Data = "data";
-        public const string Hash = "hash";
+        public const string Raw = "raw.txt";
+        public const string Data = "data.txt";
+        public const string Hash = "hash.txt";
     }
 }